Added new class, edited interface.
As discussed in class, the admin, transaction, and inventory classes
will still implement a hashmap<String, Integer> from
AuthorizationInterface. Each class will add the actions that need
authorizing to that hashmap. Those classes will grab the integer from
the hashmap after checking they action, pass it to the authorization
class, and a boolean will be returned determining if they're authorized
for that action.