Added new features.
Authorization is now more versatile, instead of a minimum auth level,
and checking auth level with a greater than or equal to expression;
Other classes can add specified authorization groups to an arraylist
that's inside the hash table. The isAuthorized method will check to see
if current user is in a group that's permitted to access that action.
Check authTest for more details and examples.